refactor(rollback): move rollback storage ops out of the StateStore trait
The 7 rollback methods (rollback_history_list/insert,
rollback_plan_collect_substates/blocks, state_tree_truncate_to_version,
substates_rewind_to_state_version, rollback_delete_after_epoch) are never
called by consensus, so they no longer pollute the StateStoreReadTransaction
and StateStoreWriteTransaction trait surface.

Logic now lives in tari_validator_rollback::storage as free functions taking
the concrete RocksDb*Transaction types. RollbackHistoryEntry stays in
tari_ootle_storage and RollbackHistoryCf stays in tari_state_store_rocksdb
since they're part of the registered DB schema; only the read/write logic
moves. The plan/stat structs (SubstateRewindPlanRow, BlocksAfterEpochRow,
RewindTransitionKind, RollbackDeleteStats, StateTreeTruncateStats,
SubstateRewindStats) move to the tool — they're function signatures, not
stored data.

Tests stay in state_store_rocksdb/tests/ (helpers live there) and call the
tool's free functions via a dev-dep on tari_validator_rollback. Cargo
permits this because dev-deps don't form a real build cycle.
