+++ crates/storage/src/consensus_models/mod.rs
@@ -17,16 +17,12 @@ mod lock_conflict;
 mod lock_intent;
 mod locked_epoch;
 mod no_vote;
-mod rollback_delete;
 mod rollback_history;
-mod rollback_plan;
 mod state_transition;
 mod state_tree_diff;
-mod state_tree_truncate;
 mod substate;
 mod substate_change;
 mod substate_lock;
-mod substate_rewind;
 mod substate_update_batch;
 mod transaction;
 mod transaction_execution;
@@ -51,16 +47,12 @@ pub use lock_conflict::*;
 pub use lock_intent::*;
 pub use locked_epoch::*;
 pub use no_vote::*;
-pub use rollback_delete::*;
 pub use rollback_history::*;
-pub use rollback_plan::*;
 pub use state_transition::*;
 pub use state_tree_diff::*;
-pub use state_tree_truncate::*;
 pub use substate::*;
 pub use substate_change::*;
 pub use substate_lock::*;
-pub use substate_rewind::*;
 pub use substate_update_batch::*;
 pub use transaction::*;
 pub use transaction_execution::*;
