+++ crates/storage/src/consensus_models/rollback_history.rs
@@ -4,15 +4,17 @@
 use serde::{Deserialize, Serialize};
 use tari_ootle_common_types::{Epoch, ShardGroup};
 
-use crate::{StateStoreReadTransaction, StateStoreWriteTransaction, StorageError};
-
-/// Breadcrumb written by the offline `tari_validator_rollback` tool after each successful
+/// Breadcrumb appended by the offline `tari_validator_rollback` tool after each successful
 /// rollback. Append-only. Records what was rolled back and when — not the full audit
 /// detail, which lives in the operator-supplied audit file.
 ///
-/// The `audit_file_basename` is the filename (no path) of the audit file the tool
-/// emitted at the same moment; operators can use it to cross-reference the archived
-/// audit with the DB's record that a rollback happened.
+/// `audit_file_basename` is the filename (no path) of the audit file the tool emitted
+/// at the same moment; operators can use it to cross-reference the archived audit with
+/// the DB's record that a rollback happened.
+///
+/// This type lives in `tari_ootle_storage` (rather than the rollback tool) only because
+/// it is the value type of `RollbackHistoryCf` registered in the rocksdb schema. Reads
+/// and writes both go through the rollback tool's `storage` module.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RollbackHistoryEntry {
     pub target_epoch: Epoch,
@@ -20,14 +22,4 @@ pub struct RollbackHistoryEntry {
     pub applied_at_unix_secs: u64,
     pub tool_version: String,
     pub audit_file_basename: String,
-}
-
-impl RollbackHistoryEntry {
-    pub fn insert<TTx: StateStoreWriteTransaction>(&self, tx: &mut TTx) -> Result<(), StorageError> {
-        tx.rollback_history_insert(self)
-    }
-
-    pub fn list<TTx: StateStoreReadTransaction>(tx: &TTx) -> Result<Vec<RollbackHistoryEntry>, StorageError> {
-        tx.rollback_history_list()
-    }
 }
