+++ utilities/validator_rollback/src/lib.rs
@@ -12,3 +12,4 @@ pub mod audit;
 pub mod audit_writer;
 pub mod convert;
 pub mod inspect;
+pub mod storage;
